% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype.data}
\alias{celltype.data}
\title{aa}
\format{
A tibble with 4 columns:
\describe{
  \item{xCell2_labels}{xCell2_labels}
  \item{all_labels}{all_labels}
  \item{ont}{ontl}
  \item{essential_genes}{essential_genes}
}
}
\usage{
celltype.data
}
\description{
Gaaa
}
\keyword{datasets}
