% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R,
%   R/xcell2object-methods.R
\name{dependencies}
\alias{dependencies}
\alias{setDeps<-}
\alias{getDeps,xCell2Object-method}
\alias{getDeps}
\alias{setDeps<-,xCell2Object-method}
\title{Access Cell Type Dependencies}
\usage{
getDeps(object)

setDeps(object) <- value

\S4method{getDeps}{xCell2Object}(object)

\S4method{setDeps}{xCell2Object}(object) <- value
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}

\item{value}{A list of hierarchical dependencies (for the setter).}
}
\value{
For `getDeps`, a list of hierarchical dependencies. 
        For `setDeps<-`, the updated \linkS4class{xCell2Object}.

For `getDeps`, a list of hierarchical dependencies. 
        For `setDeps<-`, the updated \linkS4class{xCell2Object}.
}
\description{
Retrieve or assign hierarchical dependencies between cell types for an \linkS4class{xCell2Object}.

Retrieve or assign the hierarchical dependencies between cell types for an \linkS4class{xCell2Object}.
}
\examples{
data(DICE_demo.xCell2Ref, package = "xCell2")
setDeps(DICE_demo.xCell2Ref) <- list("Parent" = "ChildType1")
getDeps(DICE_demo.xCell2Ref)
data(DICE_demo.xCell2Ref, package = "xCell2")
setDeps(DICE_demo.xCell2Ref) <- list("ParentType" = "ChildType1")
}
\seealso{
\link{xCell2Object-class}

\link{xCell2Object-class}
}
