% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DICE_demo.xCell2Ref}
\alias{DICE_demo.xCell2Ref}
\title{A demo xCell2 reference from a sub-set of the DICE reference}
\format{
An object of class \code{xCell2Object} of length 1.
}
\usage{
DICE_demo.xCell2Ref
}
\description{
This is a demo xCell2 reference object to learn how to use 'xCell2Analysis'
}
\keyword{datasets}
