% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dice_demo_ref}
\alias{dice_demo_ref}
\title{Subset of the DICE Reference}
\format{
A \linkS4class{SummarizedExperiment} object.
}
\source{
Schmiedel B et al. (2018).
}
\usage{
data(dice_demo_ref, package = "xCell2")
}
\description{
Demo reference object derived from the DICE dataset for training xCell2 references.
}
\details{
Subset of the DICE Reference

A subset of the DICE reference stored as a `SummarizedExperiment` object for the xCell 2.0 vignette.
This demo reference object demonstrates the use of \code{xCell2Train} for generating a custom xCell2 reference.
}
\seealso{
\code{\link{xCell2Train}} for generating references, and \code{\link{xCell2Analysis}} for enrichment analysis.
}
\keyword{datasets,}
\keyword{references}
