% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{getDeps}
\alias{getDeps}
\title{Get Cell Type Dependencies}
\usage{
\S4method{getDeps}{xCell2Object}(object)
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}
}
\value{
A list of cell type dependencies.
}
\description{
Retrieve hierarchical dependencies between cell types from an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
getDeps(obj)
}
