% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{getDeps}
\alias{getDeps}
\title{Get Dependencies}
\usage{
getDeps(object)
}
\arguments{
\item{object}{An \code{xCell2Object}.}
}
\value{
A list containing the cell type dependencies.
}
\description{
Access the \code{dependencies} slot from an \code{xCell2Object}.
}
\examples{
getDeps(myObject)
}
