% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{getGenesUsed}
\alias{getGenesUsed}
\title{Get Genes Used}
\usage{
\S4method{getGenesUsed}{xCell2Object}(object)
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}
}
\value{
A character vector of genes.
}
\description{
Retrieve the genes used in training the reference from an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
getGenesUsed(obj)
}
