% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{getGenesUsed}
\alias{getGenesUsed}
\title{Get Genes Used}
\usage{
getGenesUsed(object)
}
\arguments{
\item{object}{An \code{xCell2Object}.}
}
\value{
A character vector of gene names used to train the xCell2 reference object
}
\description{
Access the \code{genes_used} slot from an \code{xCell2Object}.
}
\examples{
getGenesUsed(myObject)
}
