% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{getParams}
\alias{getParams}
\title{Get Transformation Parameters}
\usage{
\S4method{getParams}{xCell2Object}(object)
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}
}
\value{
A data frame of transformation parameters.
}
\description{
Retrieve linear transformation parameters from an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
getParams(obj)
}
