% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R
\name{getParams}
\alias{getParams}
\title{Get Parameters}
\usage{
getParams(object)
}
\description{
Retrieve the linear transformation parameters for cell types from an \linkS4class{xCell2Object}.
}
