% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{getSignatures}
\alias{getSignatures}
\title{Get Cell Type Signatures}
\usage{
\S4method{getSignatures}{xCell2Object}(object)
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}
}
\value{
A list of cell type-specific gene signatures.
}
\description{
Retrieve the cell type-specific gene signatures from an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
getSignatures(obj)
}
