% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{getSignatures}
\alias{getSignatures}
\title{Get Signatures}
\usage{
getSignatures(object)
}
\arguments{
\item{object}{An \code{xCell2Object}.}
}
\value{
A list containing the xCell2 signatures.
}
\description{
Access the \code{signatures} slot from an \code{xCell2Object}.
}
\examples{
data("DICE_demo.xCell2Ref")
getSignatures(DICE_demo.xCell2Ref)
}
