% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{getSpillMat}
\alias{getSpillMat}
\title{Get Spillover Matrix}
\usage{
\S4method{getSpillMat}{xCell2Object}(object)
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}
}
\value{
A matrix of spillover correction factors.
}
\description{
Retrieve the spillover correction matrix from an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
getSpillMat(obj)
}
