% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{getSpillMat}
\alias{getSpillMat}
\title{Get Spillover Matrix}
\usage{
getSpillMat(object)
}
\arguments{
\item{object}{An \code{xCell2Object}.}
}
\value{
A matrix representing the spillover correction factors.
}
\description{
Access the \code{spill_mat} slot from an \code{xCell2Object}.
}
\examples{
data("DICE_demo.xCell2Ref")
getSpillMat(DICE_demo.xCell2Ref)
}
