% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs.genelist}
\alias{hs.genelist}
\title{List of human genes that can be removed from ref}
\format{
A data frame with 3 columns:
\describe{
  \item{gene_group}{gene groups}
  \item{ensembl}{gene Ensemble ID}
  \item{symbol}{gene symbol}
}
}
\usage{
hs.genelist
}
\description{
Genes include 7 groups: "Rb", "Mrp", "other_Rb", "chrM", "MALAT1", "chrX" and "chrY" (Credit from GitHub: BayesPrism/inst/extdata/genelist.hs.new.txt)
}
\keyword{datasets}
