% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_mouse_gene_symbols}
\alias{human_mouse_gene_symbols}
\title{Human to mouse (symbol)}
\format{
A data frame with 2 columns:
\describe{
  \item{human}{gene human}
  \item{mouse}{mouse}
}
}
\usage{
human_mouse_gene_symbols
}
\description{
A table for converting human to mouse gene symbols
}
\keyword{datasets}
