% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ImmuneCompendium.xCell2Ref}
\alias{ImmuneCompendium.xCell2Ref}
\title{An xCell2 object trained from the Immune Compendium reference}
\format{
An xCell2Object (S4) that contain:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{An list specifying cell types dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
Reference curated by Zaitsev A (2022) and trained by Angel A, et al (2024).
}
\usage{
ImmuneCompendium.xCell2Ref
}
\description{
Platform: RNA-Seq
Organism: Human
Tissue: Immune/Blood
}
\details{
You can use this reference object as in input for `xCell2Analysis` or create a new reference object using `xCell2Train`.
}
\examples{
xCell2Ref <- xCell2::ImmuneCompendium.xCell2Ref
results.xcell2 <- xCell2Analysis(mix, xcell2object = xCell2Ref)
}
\keyword{datasets}
