% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_owlready2}
\alias{install_owlready2}
\title{use doc at https://rstudio.github.io/reticulate/articles/python_dependencies.html
user support for python module}
\usage{
install_owlready2(..., envname = "r-owlready2")
}
\arguments{
\item{\dots}{as needed}

\item{envname}{character(1) arbitrary}
}
\description{
use doc at https://rstudio.github.io/reticulate/articles/python_dependencies.html
user support for python module
}
