% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mix_demo}
\alias{mix_demo}
\title{Demo Bulk Gene Expression Data}
\format{
A matrix with genes (rows) and samples (columns).
}
\usage{
data(mix_demo, package = "xCell2")
}
\description{
Example RNA-Seq data to demonstrate \code{xCell2Analysis}.
}
\details{
Demo Bulk Gene Expression Data (RNA-Seq)

A demo mixture matrix for bulk RNA-Seq gene expression data.  
Use this dataset to test \code{xCell2Analysis} with pre-trained xCell2 references.
}
\seealso{
\code{\link{xCell2Analysis}} for enrichment analysis.
}
\keyword{datasets,}
\keyword{examples}
