% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseRNAseqData.xCell2Ref}
\alias{MouseRNAseqData.xCell2Ref}
\title{Mouse RNAseq Data Reference}
\format{
An `xCell2Object` (S4) containing:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{A list specifying cell type dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
Generated by Benayoun B (2019); trained by Angel A, et al. (2024).
}
\usage{
MouseRNAseqData.xCell2Ref
}
\description{
A reference object for use in `xCell2Analysis` or to create a new reference using `xCell2Train`.
}
\details{
xCell2 Object Trained from the MouseRNAseqData Reference

An xCell2 reference object created from the MouseRNAseqData reference.
}
\examples{
xCell2Ref <- MouseRNAseqData.xCell2Ref
}
\keyword{datasets}
