% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R
\name{setDeps<-}
\alias{setDeps<-}
\title{Set Dependencies}
\usage{
setDeps(object) <- value
}
\description{
Assign the hierarchical dependencies between cell types to an \linkS4class{xCell2Object}.
}
