% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{setDeps}
\alias{setDeps}
\title{Set Cell Type Dependencies}
\usage{
\S4method{setDeps}{xCell2Object}(object) <- value
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}

\item{value}{A list of cell type dependencies.}
}
\value{
The modified \linkS4class{xCell2Object}.
}
\description{
Assign hierarchical dependencies between cell types to an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
setDeps(obj) <- list("Dependency1" = "ParentType")
}
