% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R
\name{setGenesUsed<-}
\alias{setGenesUsed<-}
\title{Set Genes Used}
\usage{
setGenesUsed(object) <- value
}
\description{
Assign the genes used in training the reference to an \linkS4class{xCell2Object}.
}
