% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{setGenesUsed}
\alias{setGenesUsed}
\title{Set Genes Used}
\usage{
\S4method{setGenesUsed}{xCell2Object}(object) <- value
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}

\item{value}{A character vector of genes.}
}
\value{
The modified \linkS4class{xCell2Object}.
}
\description{
Assign the genes used in training the reference to an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
setGenesUsed(obj) <- c("GeneA", "GeneB", "GeneC")
}
