% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{setParams<-}
\alias{setParams<-}
\title{Set Params}
\usage{
setParams(object) <- value
}
\arguments{
\item{object}{An \code{xCell2Object}.}

\item{value}{A data frame/tibble to set in the \code{params} slot.}
}
\value{
The modified \code{xCell2Object}.
}
\description{
Set the \code{params} slot in an \code{xCell2Object}.
}
\examples{
setParams(myObject) <- tibble(celltype = "B cells", a = 0.856, b = 0.197, m = 1.85, n = 0.00185)
}
