% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{setParams}
\alias{setParams}
\title{Set Transformation Parameters}
\usage{
\S4method{setParams}{xCell2Object}(object) <- value
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}

\item{value}{A data frame of transformation parameters.}
}
\value{
The modified \linkS4class{xCell2Object}.
}
\description{
Assign linear transformation parameters to an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
setParams(obj) <- data.frame(param1 = 1, param2 = 2)
}
