% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R
\name{setSignatures<-}
\alias{setSignatures<-}
\title{Set Signatures}
\usage{
setSignatures(object) <- value
}
\description{
Assign cell type-specific gene signatures to an \linkS4class{xCell2Object}.
}
