% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{setSignatures}
\alias{setSignatures}
\title{Set Cell Type Signatures}
\usage{
\S4method{setSignatures}{xCell2Object}(object) <- value
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}

\item{value}{A list of cell type-specific gene signatures.}
}
\value{
The modified \linkS4class{xCell2Object}.
}
\description{
Assign cell type-specific gene signatures to an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
setSignatures(obj) <- list("Signature1" = c("GeneA", "GeneB"))
}
