% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R
\name{setSpillMat<-}
\alias{setSpillMat<-}
\title{Set Spillover Matrix}
\usage{
setSpillMat(object) <- value
}
\description{
Assign the spillover correction matrix to an \linkS4class{xCell2Object}.
}
