% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-methods.R
\name{setSpillMat}
\alias{setSpillMat}
\title{Set Spillover Matrix}
\usage{
\S4method{setSpillMat}{xCell2Object}(object) <- value
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}

\item{value}{A matrix of spillover correction factors.}
}
\value{
The modified \linkS4class{xCell2Object}.
}
\description{
Assign the spillover correction matrix to an \linkS4class{xCell2Object}.
}
\examples{
obj <- new("xCell2Object")
setSpillMat(obj) <- matrix(0, nrow = 2, ncol = 2)
}
