% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2CleanGenes.R
\name{xCell2CleanGenes}
\alias{xCell2CleanGenes}
\title{xCell2CleanGenes function}
\usage{
xCell2CleanGenes(
  ref,
  mix,
  gene_groups = c("Rb", "Mrp", "other_Rb", "chrM", "MALAT1", "chrX", "chrY")
)
}
\arguments{
\item{ref}{Reference gene expression matrix.}

\item{mix}{Mixture gene expression matrix.}

\item{gene_groups}{Gene group to clean c("Rb", "Mrp", "other_Rb", "chrM", "MALAT1", "chrX", "chrY")}
}
\value{
A list of ref and mix with the shared genes after cleaning
}
\description{
This function clean and get shared genes between reference and mixture
}
