% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DICE_demo.xCell2Ref}
\alias{DICE_demo.xCell2Ref}
\title{Demo xCell2 Reference Object}
\format{
An `xCell2Object` (S4) containing:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{A list specifying cell type dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
Schmiedel B et al. (2018)
}
\usage{
DICE_demo.xCell2Ref
}
\description{
This object can be used to learn and demonstrate the functionality of the `xCell2Analysis` function.
}
\details{
Demo xCell2 Reference Object from a Subset of the DICE Reference

A demo xCell2 reference object derived from a subset of the DICE reference dataset.
}
\keyword{datasets}
