% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DICE_demo.xCell2Ref}
\alias{DICE_demo.xCell2Ref}
\title{Demo xCell2 Reference Object}
\format{
An \code{xCell2Object} containing:
\describe{
  \item{\code{params}}{Linear transformation parameters.}
  \item{\code{signatures}}{Cell-type-specific gene signatures.}
  \item{\code{dependencies}}{Cell type lineage dependencies.}
  \item{\code{spill_mat}}{Spillover correction matrix.}
  \item{\code{genes_used}}{Genes included in the reference.}
}
}
\source{
Schmiedel B et al. (2018).
}
\usage{
data(DICE_demo.xCell2Ref, package = "xCell2")
}
\description{
Pre-trained xCell2 reference object based on the DICE dataset.
}
\details{
Demo xCell2 Reference Object from DICE Subset (human)

A demo xCell2 reference object derived from a subset of the DICE dataset.
Suitable for demonstrating the use of \code{xCell2Analysis}.
}
\seealso{
\code{\link{xCell2Analysis}} for enrichment analysis, and \code{\link{xCell2Train}} for training custom references.
}
\keyword{datasets,}
\keyword{references}
