% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dice_demo_ref}
\alias{dice_demo_ref}
\title{A subset SummarizedExperiment object of the DICE reference for xCell 2.0 vignette}
\format{
An object of class \code{SummarizedExperiment} with 48043 rows and 60 columns.
}
\source{
Schmiedel B et al. (2018)
}
\usage{
dice_demo_ref
}
\description{
This is a demo reference object to learn how to use 'xCell2Train'
}
\keyword{datasets}
