% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dice_demo_ref}
\alias{dice_demo_ref}
\title{Subset of the DICE Reference}
\format{
A `SummarizedExperiment` object.
}
\source{
Schmiedel B et al. (2018)
}
\usage{
dice_demo_ref
}
\description{
A demo reference object derived from the DICE reference for use in the xCell 2.0 vignette.
}
\details{
A Subset of the DICE Reference

A subset `SummarizedExperiment` object of the DICE reference for the xCell 2.0 vignette.
This is a demo reference object to learn how to use the `xCell2Train` function.
}
\keyword{datasets}
