% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{getParams}
\alias{getParams}
\title{Get Params}
\usage{
getParams(object)
}
\arguments{
\item{object}{An \code{xCell2Object}.}
}
\value{
A tibble containing the linear transformation parameters.
}
\description{
Access the \code{params} slot from an \code{xCell2Object}.
}
\examples{
getParams(myObject)
}
