% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R
\name{getSignatures}
\alias{getSignatures}
\title{Get Signatures}
\usage{
getSignatures(object)
}
\description{
Retrieve the cell type-specific gene signatures from an \linkS4class{xCell2Object}.
}
