% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_mouse_gene_symbols}
\alias{human_mouse_gene_symbols}
\title{Human to Mouse Gene Symbols}
\format{
A data frame with 2 columns:
\describe{
  \item{human}{Character. Human gene symbols.}
  \item{mouse}{Character. Corresponding mouse gene symbols.}
}
}
\usage{
data(human_mouse_gene_symbols, package = "xCell2")
}
\description{
A table that provides a mapping of human gene symbols to mouse gene symbols.
}
\details{
Human to Mouse Gene Symbol Conversion Table

A data frame for converting human gene symbols to mouse gene symbols.
}
\keyword{datasets}
