% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ImmGenData.xCell2Ref}
\alias{ImmGenData.xCell2Ref}
\title{An xCell2 object trained from the Immunologic Genome Project reference}
\format{
An xCell2Object (S4) that contain:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{An list specifying cell types dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
Reference generated by The Immunological Genome Project Consortium (2008), curated by Aran D (2019) and trained by Angel A, et al (2024).
}
\usage{
ImmGenData.xCell2Ref
}
\description{
Platform: Micro-array\n
Organism: Mouse\n
Tissue: Immune/Blood
}
\details{
You can use this reference object as in input for `xCell2Analysis` or create a new reference object using `xCell2Train`.
}
\examples{
xCell2Ref <- xCell2::ImmGenData.xCell2Ref
results.xcell2 <- xCell2Analysis(mix, xcell2object = xCell2Ref)
}
\keyword{datasets}
