% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM22.xCell2Ref}
\alias{LM22.xCell2Ref}
\title{An xCell2 object trained from the LM22 reference}
\format{
An xCell2Object (S4) that contain:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{An list specifying cell types dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
Reference generated by Newman AM (2015) and trained by Angel A, et al (2024).
}
\usage{
LM22.xCell2Ref
}
\description{
Platform: Micro-array
Organism: Human
Tissue: General
}
\details{
You can use this reference object as in input for `xCell2Analysis` or create a new reference object using `xCell2Train`.
}
\examples{
xCell2Ref <- xCell2::LM22.xCell2Ref
}
\keyword{datasets}
