% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mix_demo}
\alias{mix_demo}
\title{A demo for a bulk gene expression data (RNA-Seq)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 3 columns.
}
\usage{
mix_demo
}
\description{
This is a demo mixture matrix to learn how to use 'xCell2Train'
}
\keyword{datasets}
