% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mix_demo}
\alias{mix_demo}
\title{Demo Bulk Gene Expression Data}
\format{
A matrix with rows representing genes and columns representing samples.
}
\usage{
data(mix_demo, package = "xCell2")
}
\description{
A demo mixture matrix for bulk RNA-Seq gene expression data. This dataset can be used to learn how to use the `xCell2Train` function.
}
\details{
Demo Bulk Gene Expression Data (RNA-Seq)

A demo mixture matrix for bulk RNA-Seq gene expression data, used to demonstrate how to use the `xCell2Train` function.
}
\keyword{datasets}
