% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseRNAseqData.xCell2Ref}
\alias{MouseRNAseqData.xCell2Ref}
\title{Mouse RNA-Seq Data Reference}
\format{
An \code{xCell2Object} containing:
\describe{
  \item{\code{params}}{Linear transformation parameters.}
  \item{\code{signatures}}{Gene signatures for cell types.}
  \item{\code{dependencies}}{Cell lineage dependencies.}
  \item{\code{spill_mat}}{Spillover correction matrix.}
  \item{\code{genes_used}}{Genes included in the reference.}
}
}
\source{
Benayoun B (2019); trained by Angel A, et al. (2024).
}
\usage{
data(MouseRNAseqData.xCell2Ref, package = "xCell2")
}
\description{
Pre-trained xCell2 reference for use in \code{xCell2Analysis} or extending via \code{xCell2Train}.
}
\details{
Mouse RNA-Seq Data Reference

A pre-trained xCell2 reference object based on the MouseRNAseqData dataset.
}
\seealso{
\code{\link{xCell2Analysis}} and \code{\link{xCell2Train}}.
}
\keyword{datasets}
