% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R,
%   R/xcell2object-methods.R
\name{params}
\alias{params}
\alias{setParams<-}
\alias{getParams,xCell2Object-method}
\alias{getParams}
\alias{setParams<-,xCell2Object-method}
\title{Access Transformation Parameters}
\usage{
getParams(object)

setParams(object) <- value

\S4method{getParams}{xCell2Object}(object)

\S4method{setParams}{xCell2Object}(object) <- value
}
\arguments{
\item{object}{An \linkS4class{xCell2Object}.}

\item{value}{A data frame of transformation parameters (for the setter).}
}
\value{
For `getParams`, a data frame of transformation parameters. 
        For `setParams<-`, the updated \linkS4class{xCell2Object}.

For `getParams`, a data frame of transformation parameters. 
        For `setParams<-`, the updated \linkS4class{xCell2Object}.
}
\description{
Retrieve or assign linear transformation parameters for an \linkS4class{xCell2Object}.

Retrieve or assign linear transformation parameters for cell types for an \linkS4class{xCell2Object}.
}
\seealso{
\link{xCell2Object-class}

\link{xCell2Object-class}
}
