% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{setDeps<-}
\alias{setDeps<-}
\title{Set Dependencies}
\usage{
setDeps(object) <- value
}
\arguments{
\item{object}{An \code{xCell2Object}.}

\item{value}{A list to set in the \code{dependencies} slot.}
}
\value{
The modified \code{xCell2Object}.
}
\description{
Set the \code{dependencies} slot in an \code{xCell2Object}.
}
\examples{
setDeps(myObject) <- list("T cells, CD4+" = list("descendants" = "T cells, CD4+, memory",  "ancestors" = "T cells")))
}
