% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{setGenesUsed<-}
\alias{setGenesUsed<-}
\title{Set Genes Used}
\usage{
setGenesUsed(object) <- value
}
\arguments{
\item{object}{An \code{xCell2Object}.}

\item{value}{A character vector to set in the \code{genes_used} slot.}
}
\value{
The modified \code{xCell2Object}.
}
\description{
Set the \code{genes_used} slot in an \code{xCell2Object}.
}
\examples{
setGenesUsed(myObject) <- c("GeneA", "GeneB", "GeneC")
}
