% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-generics.R
\name{setParams<-}
\alias{setParams<-}
\title{Set Parameters}
\usage{
setParams(object) <- value
}
\description{
Assign the linear transformation parameters for cell types to an \linkS4class{xCell2Object}.
}
