% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{setSignatures<-}
\alias{setSignatures<-}
\title{Set Signatures}
\usage{
setSignatures(object) <- value
}
\arguments{
\item{object}{An \code{xCell2Object}.}

\item{value}{A list to set in the \code{signatures} slot.}
}
\value{
The modified \code{xCell2Object}.
}
\description{
Set the \code{signatures} slot in an \code{xCell2Object}.
}
\examples{
setSignatures(myObject) <- list("B cell" = c("MS4A1", "IGHM", "FCRL1", "IGHD"))
}
