% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\name{setSpillMat<-}
\alias{setSpillMat<-}
\title{Set Spillover Matrix}
\usage{
setSpillMat(object) <- value
}
\arguments{
\item{object}{An \code{xCell2Object}.}

\item{value}{A matrix to set in the \code{spill_mat} slot.}
}
\value{
The modified \code{xCell2Object}.
}
\description{
Set the \code{spill_mat} slot in an \code{xCell2Object}.
}
