% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TabulaSapiensBlood.xCell2Ref}
\alias{TabulaSapiensBlood.xCell2Ref}
\title{Tabula Sapiens Blood Reference}
\format{
An `xCell2Object` (S4) containing:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{A list specifying cell type dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
The Tabula Sapiens Consortium (2022); trained by Angel A, et al. (2024).
}
\usage{
data(TabulaSapiensBlood.xCell2Ref, package = "xCell2")
}
\description{
A reference object for use in `xCell2Analysis` or to create a new reference using `xCell2Train`.
}
\details{
xCell2 Object Trained from the Tabula Sapiens Reference

An xCell2 reference object created from the Tabula Sapiens reference.
}
\keyword{datasets}
