% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMECompendium.xCell2Ref}
\alias{TMECompendium.xCell2Ref}
\title{Tumor Microenvironment Compendium Reference}
\format{
An `xCell2Object` (S4) containing:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{A list specifying cell type dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
Curated by Zaitsev A (2022) and trained by Angel A, et al. (2024).
}
\usage{
TMECompendium.xCell2Ref
}
\description{
A reference object for use in `xCell2Analysis` or to create a new reference using `xCell2Train`.
}
\details{
xCell2 Object Trained from the Tumor Microenvironment Compendium Reference

An xCell2 reference object created from the Tumor Microenvironment Compendium reference.
}
\examples{
xCell2Ref <- TMECompendium.xCell2Ref
}
\keyword{datasets}
