% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMECompendium.xCell2Ref}
\alias{TMECompendium.xCell2Ref}
\title{Tumor Microenvironment Compendium Reference}
\format{
An `xCell2Object` (S4) containing:
\describe{
  \item{\code{params}}{A tibble of parameters for linear transformation.}
  \item{\code{signatures}}{A list containing cell-type signatures.}
  \item{\code{dependencies}}{A list specifying cell type dependencies.}
  \item{\code{spill_mat}}{A matrix representing spillover correction factors.}
  \item{\code{genes_used}}{A character vector of genes used to generate the reference.}
}
}
\source{
Curated by Zaitsev A (2022) and trained by Angel A, et al. (2024).
}
\usage{
data(TMECompendium.xCell2Ref, package = "xCell2")
}
\description{
A reference object for use in `xCell2Analysis`.
}
\details{
xCell2 Object Trained from the Tumor Microenvironment Compendium Reference

An xCell2 reference object created from the Tumor Microenvironment Compendium reference.


Normalized data used to train this `xCell2` reference object can be found here:
https://science.bostongene.com/kassandra/downloads
}
\references{
Zaitsev, A., et al. (2022). Precise reconstruction of the TME using bulk RNA-seq and a machine learning algorithm trained on artificial transcriptomes. Cancer Cell, 40(8), 879-894.
}
\keyword{datasets}
