% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2Analysis.R
\name{xCell2Analysis}
\alias{xCell2Analysis}
\title{xCell2Analysis function}
\usage{
xCell2Analysis(
  bulk,
  xcell2sigs,
  min_genes_overlap = 0.8,
  spillover_alpha = 0.5
)
}
\arguments{
\item{bulk}{A matrix containing gene expression data.}

\item{xcell2sigs}{A `xCell2Signatures` object.}
}
\value{
A data frame containing the cell type enrichment for each sample in the input matrix, as estimated by xCell2.
}
\description{
This function takes a matrix of bulk gene expression data and a `xCell2Signatures` object containing a set of signatures as input. It performs downstream analysis to identify enriched cell types in the bulk sample.
}
