% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2Analysis.R
\name{xCell2Analysis}
\alias{xCell2Analysis}
\title{xCell2Analysis function}
\usage{
xCell2Analysis(
  mix,
  xcell2object,
  minSharedGenes = 0.9,
  rawScores = FALSE,
  spillover = TRUE,
  spilloverAlpha = 0.5,
  numThreads = 1
)
}
\arguments{
\item{mix}{A bulk mixture of gene expression data (genes in rows, samples in columns).}

\item{xcell2object}{An S4 object of class `xCell2Object`.}

\item{minSharedGenes}{Minimum fraction of shared genes required between the mix and the reference (default: 0.9).}

\item{rawScores}{Boolean to indicate whether to return raw enrichment scores (default: FALSE).}

\item{spillover}{Boolean - use spillover correction on the transformed enrichment scores? (default: TRUE).}

\item{spilloverAlpha}{A numeric for spillover alpha value (default: 0.5).}

\item{numThreads}{Number of threads for parallel processing (default: 1).}
}
\value{
A data frame containing the cell type enrichment for each sample in the input matrix, as estimated by xCell2.
}
\description{
This function performs cell type enrichment analysis to identify proportions of cell types in a bulk gene expression mixture.
}
\examples{
# For detailed example read xCell2 vignette.

# Load "ready to use" xCell2 reference object or generate a new one using `xCell2Train`
PanCancer.xCell2Ref <- xCell2::PanCancer.xCell2Ref

# Load bulk RNA-Seq gene expression mixture
mix_demo <- xCell2::mix_demo

# Run xCell2 cell type enrichment analysis
xcell2_res <- xCell2::xCell2Analysis(mix = mix_demo, xcell2object = PanCancer.xCell2Ref)

}
