% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2CleanGenes.R
\name{xCell2CleanGenes}
\alias{xCell2CleanGenes}
\title{xCell2CleanGenes function}
\usage{
xCell2CleanGenes(
  ref,
  mix,
  gene_groups = c("Rb", "Mrp", "other_Rb", "chrM", "MALAT1", "chrX", "chrY"),
  use_protein_coding = TRUE,
  top_var_genes = FALSE,
  n_var_genes = 5000
)
}
\arguments{
\item{ref}{Reference gene expression matrix.}

\item{mix}{Mixture gene expression matrix.}

\item{gene_groups}{Gene group to clean c("Rb", "Mrp", "other_Rb", "chrM", "MALAT1", "chrX", "chrY")}

\item{use_protein_coding}{Boolean - use only protein coding genes?}

\item{top_var_genes}{Select top variable genes using Seurat's FindVariableFeatures function (only for scRNA-seq data)}

\item{n_var_genes}{Number of top variable genes to select  (only for scRNA-seq data)}
}
\value{
A list of ref and mix with the shared genes after cleaning
}
\description{
This function clean and get shared genes between reference and mixture
}
