% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2GetLineage.R
\name{xCell2GetLineage}
\alias{xCell2GetLineage}
\title{Identify Cell Type Lineage Dependencies}
\usage{
xCell2GetLineage(labels, outFile = NULL)
}
\arguments{
\item{labels}{A data frame with the following four columns:
\itemize{
  \item \code{"ont"}: The cell type ontology ID as a character (e.g., \code{"CL:0000545"}). Set as \code{NA} if no ontology is available.
    Ontologies can be found at \href{https://www.ebi.ac.uk/ols4/ontologies/cl}{EBI Ontology Lookup Service (OLS)} or 
    by using the \link[ontologyIndex]{ontologyIndex} package.
  \item \code{"label"}: The cell type name as a character (e.g., \code{"T-helper 1 cell"}).
  \item \code{"sample"}: The sample or cell identifier, matching the column names in the reference gene expression matrix.
  \item \code{"dataset"}: The dataset or subject source for each sample. If not applicable, a constant value can be used for all samples.
}}

\item{outFile}{An optional output file name for saving the cell type dependencies as a TSV file. If specified, the file will include 
the \code{"ont"}, \code{"label"}, \code{"descendants"}, and \code{"ancestors"} columns. This file is suitable for manual inspection 
and adjustment.}
}
\value{
\itemize{
    \item If \code{outFile} is not specified, the function returns a list of cell type dependencies. Each entry in the list corresponds 
    to a cell type, with its descendants and ancestors provided as separate components.
    \item If \code{outFile} is specified, the function writes a TSV file with the cell type dependencies, providing a warning to 
    encourage manual inspection and validation.
  }
}
\description{
This function identifies cell type dependencies based on the Cell Ontology, including both
descendants and ancestors for each cell type. 
It allows users to investigate and manually inspect cell type lineage relationships in their 
reference dataset, enabling more accurate and biologically meaningful analyses with \code{xCell2}.
}
\details{
The \code{xCell2GetLineage} function enables advanced users to generate a cell type dependency file for manual inspection and adjustment. 
This file can then be passed to the \code{\link{xCell2Train}} function via the \code{lineageFile} parameter, ensuring that ontological 
integration (\code{useOntology}) operates with user-curated dependencies for maximum reliability.

If the user chooses not to use this function, \code{xCell2Train} will automatically detect cell type lineage dependencies during the 
reference training process. However, manual inspection allows users to refine lineage relationships for more precise and biologically 
meaningful cell type signatures.

If no ontologies (\code{"ont"}) are provided in the labels data frame, the function will notify the user and return a default structure 
with empty dependencies.

## Relationship with Other Function(s)
The dependencies generated by this function enhance the \code{xCell2Train} process by refining lineage-based cell type relationships. 
These refined relationships improve the quality of cell type-specific gene signatures, leading to more accurate results in 
\code{\link{xCell2Analysis}}.
}
\examples{
# For detailed example read xCell2 vignette.

library(xCell2)
 
# Load demo reference object
data(dice_demo_ref, package = "xCell2")

# Prepare labels data frame
dice_labels <- SummarizedExperiment::colData(dice_demo_ref)
dice_labels <- as.data.frame(dice_labels) # "label" column already exists
dice_labels$ont <- NA
dice_labels$sample <- colnames(dice_demo_ref)
dice_labels$dataset <- "DICE"

# Assign cell type ontology
dice_labels[dice_labels$label == "B cells", ]$ont <- "CL:0000236"
dice_labels[dice_labels$label == "Monocytes", ]$ont <- "CL:0000576"
dice_labels[dice_labels$label == "NK cells", ]$ont <- "CL:0000623"
dice_labels[dice_labels$label == "T cells, CD8+", ]$ont <- "CL:0000625"
dice_labels[dice_labels$label == "T cells, CD4+", ]$ont <- "CL:0000624"
dice_labels[dice_labels$label == "T cells, CD4+, memory", ]$ont <- "CL:0000897"

# Generate cell type lineage dependencies
xCell2::xCell2GetLineage(labels = dice_labels)

# Manually inspect and adjust the saved file "demo_dice_dep.tsv"
# Use this file as input to xCell2Train via the `lineageFile` parameter for customized
# ontological integration.
}
\seealso{
\code{\link{xCell2Train}}, for incorporating lineage relationships during reference training. \cr
\code{\link{xCell2Analysis}}, which uses the trained reference object for cell type enrichment analysis. \cr
\code{\link{AnnotationHub}}, for accessing ontology data. \cr
\code{\link{ontologyIndex}}, for exploring cell ontologies programmatically.
}
\author{
Almog Angel and Dvir Aran
}
