% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2GetLineage.R
\name{xCell2GetLineage}
\alias{xCell2GetLineage}
\title{xCell2GetLineage function}
\usage{
xCell2GetLineage(labels, out_file)
}
\arguments{
\item{labels}{A data frame containing two columns: first column named "ont" for cell type onthology (character) and second column named "label" for cell type labels (character)}

\item{out_file}{An optional output file name to write the cell type dependencies to.}
}
\value{
A list of cell type dependencies, or a TSV file containing the cell type dependencies if an output file is specified.
}
\description{
This function uses the `ontoProc` and `ontologyIndex` packages to identify the descendants and ancestors of each cell type based on the cell ontology tree. If no output file is specified, the function returns a list of cell type dependencies as a list. If an output file is specified, the function writes the cell type dependencies to a TSV file.
}
