% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2GetLineage.R
\name{xCell2GetLineage}
\alias{xCell2GetLineage}
\title{xCell2GetLineage function}
\usage{
xCell2GetLineage(labels, outFile = NULL)
}
\arguments{
\item{labels}{A data frame with four columns:
"ont": the cell type ontology as a character (i.e., "CL:0000545" or NA if there is no ontology).
"label": the cell type name as a character (i.e., "T-helper 1 cell").
"sample": the cell type sample that match the column name in ref.
"dataset": sample's source dataset or subject (can be the same for all sample if no such information).}

\item{outFile}{An optional output file name to write the cell type dependencies.}
}
\value{
A list of cell type dependencies, or a TSV file containing the cell type dependencies if an output file is specified.
}
\description{
This function identify the descendants and ancestors of each cell type based on the cell ontology tree.
If no output file is specified, the function returns a list of cell type dependencies.
If an output file is specified, the function writes the cell type dependencies to a TSV file.
}
