% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcell2object-class.R
\docType{class}
\name{xCell2Object-class}
\alias{xCell2Object-class}
\title{xCell2Object Class}
\description{
An S4 class to represent the xCell2 reference object.
}
\section{Slots}{

\describe{
\item{\code{signatures}}{List of xCell2 signatures.}

\item{\code{dependencies}}{List of cell type dependencies.}

\item{\code{params}}{Tibble of cell type linear transformation parameters.}

\item{\code{spill_mat}}{Matrix of cell types spillover correction factors.}

\item{\code{genes_used}}{Character vector of genes used to train the xCell2 reference object.}
}}

