% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2Train.R
\name{xCell2Train}
\alias{xCell2Train}
\title{xCell2Train function}
\usage{
xCell2Train(
  ref,
  mix = NULL,
  labels,
  refType,
  humanToMouse = FALSE,
  lineageFile = NULL,
  seed = NULL,
  numThreads = 1,
  useOntology = TRUE,
  returnSignatures = FALSE,
  returnAnalysis = FALSE,
  useSpillover = TRUE,
  spilloverAlpha = 0.5,
  minPbCells = 30,
  minPbSamples = 10,
  minScGenes = 10000,
  topSpillValue = 0.5
)
}
\arguments{
\item{ref}{A reference gene expression matrix (genes in rows, samples/cells in columns).}

\item{mix}{A bulk mixture of gene expression data (genes in rows, samples in columns) (optional).}

\item{labels}{A data frame in which the rows correspond to samples in the ref. The data frame must have four columns:
"ont": the cell type ontology as a character (i.e., "CL:0000545" or NA if there is no ontology).
"label": the cell type name as a character (i.e., "T-helper 1 cell").
"sample": the cell type sample/cell that match the column name in ref.
"dataset": sample's source dataset or subject (can be the same for all samples if no such information).}

\item{refType}{The reference gene expression data type: "rnaseq" for bulk RNA-Seq, "array" for micro-array, or "sc" for scRNA-Seq.}

\item{humanToMouse}{A Boolean for converting human genes to mouse genes (default: FALSE).}

\item{lineageFile}{Path to the cell type lineage file generated with `xCell2GetLineage` function and reviewed manually (optional).}

\item{seed}{Set seed for reproducible results (default: NULL).}

\item{numThreads}{Number of threads for parallel processing (default: 1).}

\item{useOntology}{A Boolean for considering cell type dependencies by using ontological integration (default: TRUE).}

\item{returnSignatures}{A Boolean to return just the signatures (default: FALSE).}

\item{returnAnalysis}{A Boolean to return the xCell2Analysis results (do not return reference object) (default: FALSE).}

\item{useSpillover}{A Boolean to use spillover correction in xCell2Analysis (returnAnalysis must be TRUE) (default: TRUE).}

\item{spilloverAlpha}{A numeric for spillover alpha value in xCell2Analysis (returnAnalysis must be TRUE) (default: 0.5).}

\item{minPbCells}{For scRNA-Seq reference only - minimum number of cells in the pseudo-bulk (optional, default: 30).}

\item{minPbSamples}{For scRNA-Seq reference only - minimum number of pseudo-bulk samples (optional, default: 10).}

\item{minScGenes}{For scRNA-Seq reference only - minimum number of genes for pseudo-bulk samples (default: 10000).}

\item{topSpillValue}{Maximum spillover compensation correction value (default: 0.5).}
}
\value{
An S4 object containing cell types' signatures, linear transformation parameters, spillover matrix and dependencies.
}
\description{
This function generates a custom xCell2 reference object for cell type enrichment analysis.
}
\examples{
# For detailed example read xCell2 vignette.

# Extract reference matrix
dice <- xCell2::dice_demo_ref
dice_ref <- as.matrix(dice@assays@data$logcounts)
colnames(dice_ref) <- make.unique(colnames(dice_ref)) # Make samples samples unique

# Extract reference metadata
dice_labels <- as.data.frame(dice@colData)

# Prepare labels data frame
dice_labels$ont <- NA
dice_labels$sample <- colnames(dice_ref)
dice_labels$dataset <- "DICE"

# Assign cell type ontology (optional but recommended)
dice_labels[dice_labels$label == "B cells",]$ont <- "CL:0000236"
dice_labels[dice_labels$label == "Monocytes",]$ont <- "CL:0000576"
dice_labels[dice_labels$label == "NK cells",]$ont <- "CL:0000623"
dice_labels[dice_labels$label == "T cells, CD8+",]$ont <- "CL:0000625"
dice_labels[dice_labels$label == "T cells, CD4+",]$ont <- "CL:0000624"
dice_labels[dice_labels$label == "T cells, CD4+, memory",]$ont <- "CL:0000897"

# Generate custom xCell2 reference object
DICE.xCell2Ref <- xCell2::xCell2Train(ref = dice_ref, labels = dice_labels, refType = "rnaseq")

}
