% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2Train.R
\name{xCell2Train}
\alias{xCell2Train}
\title{xCell2Train function}
\usage{
xCell2Train(
  ref,
  labels,
  data_type,
  mix = NULL,
  lineage_file = NULL,
  top_genes_frac = 0.5,
  medianGEP = TRUE,
  seed = 123,
  probs = c(0.01, 0.05, 0.1, 0.25, 0.333, 0.49),
  sim_fracs = c(0, seq(0.01, 0.25, 0.005), seq(0.3, 1, 0.05)),
  diff_vals = c(1, 1.585, 2, 3, 4, 5),
  min_genes = 3,
  max_genes = 100,
  return_sigs = FALSE,
  sigsFile = NULL,
  minPBcells = 30,
  minPBsamples = 10,
  ct_sims = 20,
  samples_frac = 0.1,
  simMethod = "ref_thin",
  nCores = 1
)
}
\arguments{
\item{ref}{A reference gene expression matrix.}

\item{labels}{A data frame in which the rows correspond to samples in the ref. The data frame must have four columns:
"ont": the cell type ontology as a character (i.e., "CL:0000545" or NA if there is no ontology).
"label": the cell type name as a character (i.e., "T-helper 1 cell").
"sample": the cell type sample should match the column name in ref.
"dataset": the cell type sample dataset or subject (for single-cell) as a character.}

\item{data_type}{Gene expression data type: "rnaseq", "array", or "sc".}

\item{mix}{description}

\item{lineage_file}{Path to the cell type lineage file generated with `xCell2GetLineage` function (optional).}

\item{top_genes_frac}{description}

\item{probs}{A vector of probability thresholds to be used for generating signatures (optional).}

\item{sim_fracs}{A vector of mixture fractions to be used in signature filtering (optional).}

\item{diff_vals}{A vector of delta values to be used for generating signatures (optional).}

\item{min_genes}{The minimum number of genes to include in the signature (optional).}

\item{max_genes}{The maximum number of genes to include in the signature (optional).}

\item{return_sigs}{description}

\item{sigsFile}{description}

\item{minPBcells}{description}

\item{ct_sims}{description}

\item{samples_frac}{description}

\item{simMethod}{description}

\item{nCores}{description}

\item{minPBgroups}{description}
}
\value{
An S4 object containing the signatures, cell type labels, and cell type dependencies.
}
\description{
This function generates signatures for each cell type.
}
