% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCell2Train.R
\name{xCell2Train}
\alias{xCell2Train}
\title{xCell2Train function}
\usage{
xCell2Train(
  ref,
  labels,
  data_type,
  lineage_file = NULL,
  weightGenes = FALSE,
  sim_fracs = c(0, 0.001, 0.002, 0.004, 0.006, 0.008, seq(0.01, 1, 0.01)),
  diff_vals = c(1, 1.32, 1.585, 2, 3, 4, 5),
  min_genes = 5,
  max_genes = 200,
  filter_sigs = TRUE,
  sigsFile = NULL,
  params = list(),
  modelType = "rf",
  bulkPseudoCount = 1,
  minPBcells = 30,
  minPBgroups = 10
)
}
\arguments{
\item{ref}{A reference gene expression matrix.}

\item{labels}{A data frame in which the rows correspond to samples in the ref. The data frame must have four columns:
"ont": the cell type ontology as a character (i.e., "CL:0000545" or NA if there is no ontology).
"label": the cell type name as a character (i.e., "T-helper 1 cell").
"sample": the cell type sample should match the column name in ref.
"dataset": the cell type sample dataset or subject (for single-cell) as a character.}

\item{data_type}{Gene expression data type: "rnaseq", "array", or "sc".}

\item{lineage_file}{Path to the cell type lineage file generated with `xCell2GetLineage` function (optional).}

\item{weightGenes}{description}

\item{sim_fracs}{A vector of mixture fractions to be used in signature filtering (optional).}

\item{diff_vals}{A vector of delta values to be used for generating signatures (optional).}

\item{min_genes}{The minimum number of genes to include in the signature (optional).}

\item{max_genes}{The maximum number of genes to include in the signature (optional).}

\item{filter_sigs}{description}

\item{sigsFile}{description}

\item{params}{description}

\item{modelType}{description}

\item{bulkPseudoCount}{description}

\item{minPBcells}{description}

\item{minPBgroups}{description}

\item{probs}{A vector of probability thresholds to be used for generating signatures (optional).}
}
\value{
An S4 object containing the signatures, cell type labels, and cell type dependencies.
}
\description{
This function generates signatures for each cell type.
}
